Update to raylib 3.7
Raylib 3.7 significantly speeds the "thingy" example by 2x.

Additionally the raylib that was linked against was actually
compiled with Zig (zig cc) instead of gcc. This fixes a linking
error when using optimizations greater that -O0. (The system libc
and the one provided by llvm had some issues with vfprintf)